+++ frontend/src/app/enums/http-request-method.enum.ts
@@ -1,0 +1,4 @@
+export enum HttpRequestMethod {
+    GET = 'GET',
+    POST = 'POST',
+}