+++ frontend/src/app/features/main-page/components/add-categories-form/add-categories-form.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'mados-add-categories-form',
+  templateUrl: './add-categories-form.component.html',
+  styleUrls: ['./add-categories-form.component.scss']
+})
+export class AddCategoriesFormComponent { }