+++ frontend/src/app/features/main-page/components/add-game-form/add-game-form.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+
+import { AddGameFormComponent } from './add-game-form.component';
+
+describe('AddGameFormComponent', () => {
+  let component: AddGameFormComponent;
+  let fixture: ComponentFixture<AddGameFormComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [ AddGameFormComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(AddGameFormComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+    it('Should create a component', () => {
+        expect(component).withContext('Component should be created').toBeTruthy();
+    });
+});