+++ frontend/src/app/features/main-page/components/add-game-form/add-game-form.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'mados-add-game-form',
+  templateUrl: './add-game-form.component.html',
+  styleUrls: ['./add-game-form.component.scss']
+})
+export class AddGameFormComponent { }