+++ frontend/src/app/features/main-page/components/description-section/components/game/game.component.ts
@@ -1,0 +1,15 @@
+import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'mados-game',
+  templateUrl: './game.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class GameComponent {
+    @Input()
+    public oneHundredPercentIconName: string;
+    @Input()
+    public name: string;
+    @Input()
+    public iconNames: string[] | null;
+}