+++ frontend/src/app/interfaces/category.interface.ts
@@ -1,0 +1,8 @@
+import { CategoryType } from '../enums/category-type.enum';
+
+export interface Category {
+    type: CategoryType;
+    iconName: string;
+    description: string | null;
+    amount: number;
+}