+++ frontend/src/app/interfaces/description-response-body.interface.ts
@@ -1,0 +1,8 @@
+import { Category } from './category.interface';
+import { Game } from './game.interface';
+
+export interface DescriptionResponseBody {
+    categories: Category[];
+    completedGames: Game[];
+    gamesWithNewAchievements: Game[];
+}