+++ frontend/src/app/interfaces/game.interface.ts
@@ -1,0 +1,5 @@
+export interface Game {
+    oneHundredPercentIconName: string;
+    name: string;
+    iconNames: string[] | null;
+}