+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Both frontend and backend changes and the final version of the application will be documented in this file.
 
+## 0.1.0 - 26.01.2022 [FE]
+
+- Add main-page module with lazy loading
+- Add various components and utils
+- Add description component
+- Add mocks
+- Add necessary enums and interfaces
+
 ## 0.0.0 - 26.01.2022 [FE]
 
 - Init frontend project